[fix] the first building commit of 2024

+++ src/ui/listbox/ui_listbox.cpp
@@ -29,7 +29,7 @@ ekg::ui::listbox *ekg::ui::listbox::set_width(float w) {
   if (this->sync_ui.w != w) {
     this->sync_ui.w = w;
 
-    ekg_bitwise_add(this->sync_flags, ekg::ui_sync::dimension);
+    ekg_bitwise_add(this->sync_flags, static_cast<uint16_t>(ekg::ui_sync::dimension));
     ekg::reload(this->id);
     ekg::synclayout(this->parent_id);
   }
